tracing/requirements.txt
azure-ai-inference[opentelemetry]
python-dotenv
azure-monitor-opentelemetry
azure-identity